+++ src/components/Wizard/Presets/Confirmation/Confirmation.scss
