+++ src/components/Wizard/Presets/Confirmation/Confirmation.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 
 import classNames from 'classnames/bind';
 
-import { Buttons, ButtonsProps } from './Buttons';
+import { Buttons, ButtonsProps } from '../Buttons';
 
 import './Confirmation.scss';
