+++ src/components/Wizard/Presets/Confirmation/index.ts
@@ -1,0 +1,1 @@
+export * from './Confirmation';
