+++ src/components/Wizard/Presets/Loading/Loading.scss
