+++ src/components/Wizard/Presets/Loading/Loading.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { LoadingIndicator } from '../../LoadingIndicator';
+import { LoadingIndicator } from '../../../LoadingIndicator';
 
 import './Loading.scss';
