+++ src/components/Wizard/Presets/NFTDetails/NFTDetails.module.scss
