+++ src/components/Wizard/Presets/NFTDetails/NFTDetails.tsx
