+++ src/components/Wizard/Presets/NFTDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './NFTDetails';
